Destroy and RequireDestroy
These are intended primarily as an easy way to clean up after a test run.
If the destroy fails then remote objects might still exist, so
RequireDestroy includes some additional logging to warn about the
consequences of that to prompt developers to check that none of the
leftover objects are billable in the remote system.