allow setting TFTEST_LOG_PATH in env
If TFTEST_LOG_PATH is set, it will be passed as TF_LOG_PATH to all Terraform commands run by tftest, with a log level of TRACE.

This is an easy and quick debugging method for use during tftest development - sensible debug logging in testing.Verbose mode is not yet implemented.

This commit also fixes a bug where the OS env was read but not passed to Terraform commands.