copy provider source code into baseDir
Provider acceptance tests sometimes read test fixture files located variously in the provider source code directory tree. Since there is not yet a naming convention for these files, we allow consumers of this library to pass in the entire source directory.

Note that the directory copied will be that containing provider_test.go, which is generally a subdirectory ("provider/") not the root of the provider source code tree.

The contents of the source directory are copied into the base directory when a new WorkingDir is created.