Add support for debugging binary acceptance testing.
Add environment variables as a map to WorkingDir, so we can associate
environment variables with the WorkingDir that will be used when running
commands using that WorkingDir.

Move getTerraformEnv to be a method on WorkingDir, so it can read the
environment variables set on the WorkingDir.

Stop propagating the TF_PLUGIN_MAGIC_COOKIE value to Terraform, it
should never be set when running terraform.

Add Setenv and Unsetenv to WorkingDir to set and unset environment
variables on the WorkingDir.

Add a GetHelper method to WorkingDir, to return the Helper associated
with the WorkingDir. We need this to get access to GetPluginName in the
SDK.

Pass the config directory when calling terraform init.